% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{compute_mcse}
\alias{compute_mcse}
\title{Compute MCSE}
\usage{
compute_mcse(error_vec, nhours = 500, n.pred = 100)
}
\arguments{
\item{error_vec}{The vector composed of sum squared prediction error}

\item{nhours}{Number of simulation times}

\item{n.pred}{Number of unmonitored locations}
}
\value{
Monte Carlo Standard Error during simulation for nhours times
}
\description{
Compute MCSE
}
