% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{melding}
\alias{melding}
\title{Title}
\usage{
melding(
  m1,
  m2,
  nm,
  sam.sloc,
  sam.sloc1,
  ton,
  burnin,
  zhat,
  Zbtilde,
  degree,
  cov.model
)
}
\arguments{
\item{cov.model}{}
}
\description{
Title
}
