% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{point_plt}
\alias{point_plt}
\title{Generate plots to show the locations of monitored/unmonitored stations when grid cell = 2}
\usage{
point_plt(resu)
}
\arguments{
\item{resu}{The output from the melding function when when grid cell = 2}
}
\value{
A ggplot with locations of monitored/unmonitored stations when grid cell = 2
}
\description{
Generate plots to show the locations of monitored/unmonitored stations when grid cell = 2
}
\examples{
point_plt(compute_param_error())
}
