% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{traceplot_plt}
\alias{traceplot_plt}
\title{Make the traceplot for all of the estimated model parameters}
\usage{
traceplot_plt(result, num_data_set)
}
\arguments{
\item{result}{The first element from the output of the melding function}

\item{num_data_set}{The number of independent during simulation, take the number from 1 to nhours}
}
\value{
Traceplot of all model parameters
}
\description{
Make the traceplot for all of the estimated model parameters
}
\examples{
traceplot_plt(compute_param_error(), 1)
}
