% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updatezs}
\alias{updatezs}
\title{Generate MCMC sample from the full conditional distribution of true underlying process Z}
\usage{
updatezs(
  X,
  zhat,
  Zbtilde,
  A1,
  A2,
  beta,
  a,
  b,
  nm,
  m2,
  sigmad,
  sigmae,
  sigma3.solve
)
}
\arguments{
\item{X}{The covariate matrix at monitored stations and sampling points within grid cells.}

\item{zhat}{The measurements vector.}

\item{Zbtilde}{The modeling outputs vector.}

\item{A1}{Matrix A1}

\item{A2}{Matrix A2}

\item{beta}{Regression coefficient}

\item{a}{Additive bias parameter a}

\item{b}{Nultiplicative bias parameter b}

\item{nm}{Number of sampling points in each grid cell of the modeling output;}

\item{m2}{Number of grid cells.}

\item{sigmad}{The output error variance parameter sigma_delta^2}

\item{sigmae}{Measurement error variance parameter sigma_e^2}

\item{sigma3.solve}{Inverse of the spatial covariance matrix of true underlying process}
}
\value{
MCMC sample of true underlying process Z
}
\description{
Generate MCMC sample from the full conditional distribution of true underlying process Z
}
