% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{melding}
\alias{melding}
\title{Run BM model and return a list of outputs}
\usage{
melding(
  m1,
  m2,
  nm,
  sam.sloc,
  sam.sloc1,
  ton,
  burnin,
  zhat,
  Zbtilde,
  degree,
  cov.model
)
}
\arguments{
\item{m1}{Number of monitoring locations}

\item{m2}{Number of grid cells}

\item{nm}{Number of points with in each of grid cells}

\item{sam.sloc}{The coordinates of sampling points and the monitoring stations}

\item{sam.sloc1}{The coordinates of the unmonitored stations where the measurements for prediction}

\item{ton}{Number of MCMC iterations in the Gibbs sampling}

\item{burnin}{The “burn-in” period of the Gibbs sampling.}

\item{zhat}{The measurements vector.}

\item{Zbtilde}{The modeling outputs vector}

\item{degree}{Degree of the polynomial function, 0 the mean is assumed constant across space,
1 the mean is assumed to be a first order polynomial on the coordinates,
2 the mean is assumed to be a second order polynomial on the coordinates.}

\item{cov.model}{one of the three possible choices for the covariance function: “exponential”, “Gaussian” and “matern”.}
}
\value{
A list of estimated model parameters
}
\description{
Run BM model and return a list of outputs
}
