% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_func}
\alias{run_func}
\title{Title}
\usage{
run_func(
  m1 = 20,
  m2 = 2,
  n.pred = 100,
  ran = 5,
  nm = 1,
  nhours = 50,
  ton = 1000,
  burnin = ton/10,
  degree = 1,
  nbeta = 3,
  ntheta = 2,
  nab = 2,
  nerror = 2,
  cov.model = "exponential",
  seed,
  num_for_cores = 6
)
}
\arguments{
\item{num_for_cores}{}
}
\description{
Title
}
