% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simudata.R
\name{simudata_data_position}
\alias{simudata_data_position}
\title{Simulate the locations of monitoring/unmonitored locations and sampling points within grid cells}
\usage{
simudata_data_position(m1, m2, n.pred, ran, nm, seed = 1234)
}
\arguments{
\item{m1}{Number of monitoring locations}

\item{m2}{Number of grid cells}

\item{n.pred}{Number of unmonitored locations for prediction}

\item{ran}{Range for the monitoring/unmonitored locations and sampling points within grid cells. Lower bound is -ran and upper bound is ran}

\item{nm}{Number of points with in each of grid cells}

\item{seed}{Random seed number}
}
\value{
A list of locations: monitoring locations, grid cell locations, monitoring locations and grid cell locations, unmonitored locations.
}
\description{
Simulate the locations of monitoring/unmonitored locations and sampling points within grid cells
}
\examples{
simudata_data_position(20, 2, 100, 5, 1, 1234)
}
