% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simudata.R
\name{simudata_data_position}
\alias{simudata_data_position}
\title{Title}
\usage{
simudata_data_position(m1, m2, n.pred, ran, nm, seed = 1234)
}
\arguments{
\item{n.pred}{Number of unmonitored locations for prediction}

\item{seed}{}
}
\description{
Title
}
