% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updatesigma}
\alias{updatesigma}
\title{Generate MCMC sample from the full conditional distribution of sigma_delta^2 and sigma_e^2}
\usage{
updatesigma(zhat, Zbtilde, nm, n, m2, y, a, b, A2)
}
\arguments{
\item{zhat}{The measurements vector.}

\item{Zbtilde}{The modeling outputs vector.}

\item{nm}{Number of sampling points in each grid cell of the modeling output;}

\item{n}{Number of monitored stations and sampling points within grid cells;}

\item{m2}{Number of grid cells.}

\item{y}{The realizations of the true underlying process at monitored
stations and sampling points within grid cells.}

\item{a}{Additive bias parameter a}

\item{b}{Multiplicative bias parameter b}

\item{A2}{Matrix A2,}
}
\value{
MCMC sample of sigma
}
\description{
Generate MCMC sample from the full conditional distribution of sigma_delta^2 and sigma_e^2
}
