% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updatezs}
\alias{updatezs}
\title{Title}
\usage{
updatezs(
  X,
  zhat,
  Zbtilde,
  A1,
  A2,
  beta,
  a,
  b,
  nm,
  m2,
  sigmad,
  sigmae,
  sigma3.solve
)
}
\arguments{
\item{sigma3.solve}{}
}
\description{
Title
}
